requirements.txt
aiofiles==24.1.0
aiXplain==0.2.31
alembic==1.16.3
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cssselect2==0.8.0
dataclasses-json==0.6.7
fastapi==0.115.13
filetype==1.2.0
fonttools==4.58.5
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
h11==0.16.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
Markdown==3.8.2
MarkupSafe==3.0.2
marshmallow==3.26.1
mypy_extensions==1.1.0
numpy==2.3.1
oauthlib==3.3.1
packaging==25.0
pandas==2.3.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
sentry-sdk==2.31.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.3
validators==0.35.0
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
google-api-python-client
google-auth-oauthlib